Forked for NDK 10d build.
Updated make files to now grab the latest NDK from google.
With this new NDK version there were two main changes that needed to be made.
pytime.c - We need to undef HAVE_FTIME, did this by changed the undef to include if we are on Android
posixmodule.c - AT_EACCESS  is undefined. HAVE_FACCESSAT must be new. Not sure if commenting out the code like I did, or undef it is the proper route. I tried to undef, but it still has the issue.  Something else may be defining it.